+++ attention-comparison-batch2-head4-d32-fwd.csv
@@ -1,8 +1,8 @@
 len,Triton-Attn-Bias (FLOPS),PyTorch-Attn-Bias (FLOPS),PyTorch-SDPA (FLOPS)
-256.000000,6.188271,1.646546,3.383988
-512.000000,14.740146,2.922090,7.482720
-1024.000000,26.877395,2.817934,14.061347
-2048.000000,37.042961,2.397494,19.714920
-4096.000000,36.803839,2.762667,18.709737
-8192.000000,35.036906,3.077841,18.240496
-16384.000000,32.332642,3.205471,18.075474
+256.000000,6.044991,1.645394,3.315631
+512.000000,15.044760,2.929594,7.544028
+1024.000000,26.814166,2.819278,14.080586
+2048.000000,37.142679,2.396055,19.470077
+4096.000000,36.455769,2.763378,18.722902
+8192.000000,33.088981,3.079588,18.163183
+16384.000000,32.292442,3.193105,18.045934
