+++ flash_attn_w_bias.py
@@ -9,9 +9,8 @@ It supports
 - Gradients for the bias term are also implemented.
 
 
-
 Caution:
-- This is an *experimental* implementation; I can't guarantee it's universally correct. 
+- This is an *experimental* implementation; I can't guarantee it's universally correct.
 - Currently, the implementation is tested on A100 with common shapes.
 - Autotune can significantly improve the performance but likely increases the errors and even introduces race conditions.
 - By default we enable autotune.
@@ -55,13 +54,7 @@ import torch
 import triton
 import triton.language as tl
 
-configs = [
-    triton.Config({'BLOCK_M': BM, 'BLOCK_N': BN}, num_stages=s, num_warps=w) \
-    for BM in [32, 64, 128] \
-    for BN in [32, 64, 128] \
-    for s in ([1, 2]) \
-    for w in [4, 8] \
-    ]
+from config import configs
 
 
 def keep(conf):
@@ -77,7 +70,7 @@ def keep(conf):
 # Disabling autotune for now, set num_warps=4 if headdim=64 and num_warps=8 if headdim=128
 @triton.autotune(
     list(filter(keep, configs)),
-    key=['CACHE_KEY_SEQLEN_Q', 'CACHE_KEY_SEQLEN_K', 'BIAS_TYPE', 'IS_CAUSAL', 'BLOCK_HEADDIM']
+    key=["CACHE_KEY_SEQLEN_Q", "CACHE_KEY_SEQLEN_K", "BIAS_TYPE", "IS_CAUSAL", "BLOCK_HEADDIM"]
 )
 @triton.heuristics(
     {
@@ -444,7 +437,7 @@ def _bwd_kernel_one_col_block(
     dv = tl.zeros([BLOCK_N, BLOCK_HEADDIM], dtype=tl.float32)
     dk = tl.zeros([BLOCK_N, BLOCK_HEADDIM], dtype=tl.float32)
     # There seems to be some problem with Triton pipelining that makes results wrong for
-    # headdim=64, seqlen=(113, 255), bias_type='matrix'. In this case the for loop
+    # headdim=64, seqlen=(113, 255), bias_type="matrix". In this case the for loop
     # may have zero step, and pipelining with the bias matrix could screw it up.
     # So we just exit early.
     if begin_m >= seqlen_q:
@@ -582,7 +575,7 @@ def _bwd_kernel_one_col_block(
         ds = (db * softmax_scale).to(q.dtype)
         # compute db is ds, 
         if BIAS_TYPE != "none":
-            tl.static_assert(BIAS_TYPE == "matrix", "BIAS_TYPE must be 'matrix' for bwd bias gradient")
+            tl.static_assert(BIAS_TYPE == "matrix", "BIAS_TYPE must be \"matrix\" for bwd bias gradient")
             if EVEN_M & EVEN_N:
                 tl.store(db_ptrs, db, eviction_policy="evict_last")
             else:
@@ -599,7 +592,7 @@ def _bwd_kernel_one_col_block(
         # compute dq
         if not (
                 EVEN_M & EVEN_HEADDIM
-        ):  # Otherewise there's a race condition when BIAS_TYPE='matrix'
+        ):  # Otherwise there's a race condition when BIAS_TYPE="matrix"
             tl.debug_barrier()
         if not ATOMIC_ADD:
             if EVEN_M & EVEN_HEADDIM:  # Race condition if we just do EVEN_M
@@ -702,10 +695,10 @@ def init_to_zero(name):
         # ),
         # Other configs seem to give wrong results when seqlen_q % 128 != 0, disabling them for now
         # # Kernel is buggy (give wrong result) if we set BLOCK_m=128, BLOCK_n=64, num_warps=*4*
-        # triton.Config({"BLOCK_M": 128, "BLOCK_N": 64, "SEQUENCE_PARALLEL": False}, num_warps=8, num_stages=1, pre_hook=init_to_zero('DQ')),
-        # triton.Config({"BLOCK_M": 128, "BLOCK_N": 64, "SEQUENCE_PARALLEL": True}, num_warps=8, num_stages=1, pre_hook=init_to_zero('DQ')),
-        # triton.Config({"BLOCK_M": 64, "BLOCK_N": 64, "SEQUENCE_PARALLEL": False}, num_warps=4, num_stages=1, pre_hook=init_to_zero('DQ')),
-        # triton.Config({"BLOCK_M": 64, "BLOCK_N": 64, "SEQUENCE_PARALLEL": True}, num_warps=4, num_stages=1, pre_hook=init_to_zero('DQ')),
+        # triton.Config({"BLOCK_M": 128, "BLOCK_N": 64, "SEQUENCE_PARALLEL": False}, num_warps=8, num_stages=1, pre_hook=init_to_zero("DQ")),
+        # triton.Config({"BLOCK_M": 128, "BLOCK_N": 64, "SEQUENCE_PARALLEL": True}, num_warps=8, num_stages=1, pre_hook=init_to_zero("DQ")),
+        # triton.Config({"BLOCK_M": 64, "BLOCK_N": 64, "SEQUENCE_PARALLEL": False}, num_warps=4, num_stages=1, pre_hook=init_to_zero("DQ")),
+        # triton.Config({"BLOCK_M": 64, "BLOCK_N": 64, "SEQUENCE_PARALLEL": True}, num_warps=4, num_stages=1, pre_hook=init_to_zero("DQ")),
     ],
     key=["CACHE_KEY_SEQLEN_Q", "CACHE_KEY_SEQLEN_K", "BIAS_TYPE", "IS_CAUSAL", "BLOCK_HEADDIM"],
 )
@@ -1117,301 +1110,3 @@ class FlashAttnFunc(torch.autograd.Function):
                 softmax_scale=ctx.softmax_scale,
             )
         return dq, dk, dv, db, None, None
-
-
-flash_attn_w_bias = FlashAttnFunc.apply
-
-
-def attention_torch(q, k, v, softmax_scale, bias=None):
-    """
-    Implements standard scaled dot-product attention using PyTorch.
-
-    Args:
-        q: Query tensor of shape (batch_size, seqlen_q, nheads, headdim)
-        k: Key tensor of shape (batch_size, seqlen_k, nheads, headdim)
-        v: Value tensor of shape (batch_size, seqlen_k, nheads, headdim)
-        softmax_scale: Scalar for softmax normalization (default: 1/sqrt(headdim))
-        bias: Optional bias tensor, shape (batch, nheads, seqlen_q, seqlen_k)
-
-    Returns:
-        Output tensor of shape (batch_size, seqlen_q, nheads, headdim)
-    """
-    # Compute scaled dot-product attention scores
-    attn_scores = torch.einsum("bqnh,bknh->bqnk", q, k)  # (batch, seqlen_q, nheads, seqlen_k)
-    # Apply softmax scaling
-    attn_scores *= softmax_scale
-
-    # Apply bias if provided
-    if bias is not None:
-        attn_scores += bias.permute(0, 2, 1, 3)
-
-    # Apply softmax along key dimension
-    attn_probs = torch.softmax(attn_scores, dim=-1)
-
-    # Compute output by weighted sum of values
-    output = torch.einsum("bqnk,bknh->bqnh", attn_probs, v)
-
-    return output
-
-
-def attention_sdpa(q, k, v, softmax_scale, bias=None):
-    """
-    Implements scaled dot-product attention using PyTorch's F.scaled_dot_product_attention.
-    
-    Args:
-        q: Query tensor of shape (batch_size, seqlen_q, nheads, headdim)
-        k: Key tensor of shape (batch_size, seqlen_k, nheads, headdim)
-        v: Value tensor of shape (batch_size, seqlen_k, nheads, headdim)
-        softmax_scale: Scalar for softmax normalization (default: 1/sqrt(headdim))
-        bias: Optional bias tensor, shape (batch, nheads, seqlen_q, seqlen_k)
-        
-    Returns:
-        Output tensor of shape (batch_size, seqlen_q, nheads, headdim)
-    """
-    # SDPA expects input of shape (B, Nh, L, D)
-    # Our inputs are already in this format after transposing
-    q = q.transpose(1, 2)  # (batch, nheads, seqlen_q, headdim)
-    k = k.transpose(1, 2)  # (batch, nheads, seqlen_k, headdim)
-    v = v.transpose(1, 2)  # (batch, nheads, seqlen_k, headdim)
-
-    # Prepare attention mask from bias
-    attn_mask = bias if bias is not None else None
-
-    # Call SDPA with our inputs
-    output = torch.nn.functional.scaled_dot_product_attention(
-        query=q,
-        key=k,
-        value=v,
-        attn_mask=attn_mask,
-        dropout_p=0.0,
-        scale=softmax_scale,
-        is_causal=False
-    )
-
-    # Return to original format (B, Lq, Nh, D)
-    return output.transpose(1, 2)
-
-
-# Benchmarking configuration
-BATCH, N_HEADS, HEAD_DIM = 2, 4, 32  # Aligned with test_attention_correctness
-
-configs = [
-    triton.testing.Benchmark(
-        x_names=["len"],
-        x_vals=list(2 ** i for i in range(8, 15)),
-        line_arg="provider",
-        line_vals=["triton-attn-bias", "torch-attn-bias", "torch-sdpa"],
-        line_names=["Triton-Attn-Bias (FLOPS)", "PyTorch-Attn-Bias (FLOPS)", "PyTorch-SDPA (FLOPS)"],
-        styles=[("red", "-"), ("blue", "--"), ("green", ":")],
-        ylabel="TFLOPS",  # Changed from Time (ms) to TFLOPS
-        plot_name=f"attention-comparison-batch{BATCH}-head{N_HEADS}-d{HEAD_DIM}-{mode}",
-        args={
-            "H": N_HEADS,
-            "BATCH": BATCH,
-            "HEAD_DIM": HEAD_DIM,
-            "mode": mode,
-        }
-    )
-    for mode in ["fwd", "bwd"]
-]
-
-
-@triton.testing.perf_report(configs)
-def bench_attention_comparison(BATCH, H, len, HEAD_DIM, mode, provider, device="cuda"):
-    dtype = torch.float16
-
-    # Initialize tensors with same shape ordering as test_attention_correctness
-    q = torch.randn((BATCH, len, H, HEAD_DIM), dtype=dtype, device=device, requires_grad=True)
-    k = torch.randn((BATCH, len, H, HEAD_DIM), dtype=dtype, device=device, requires_grad=True)
-    v = torch.randn((BATCH, len, H, HEAD_DIM), dtype=dtype, device=device, requires_grad=True)
-    bias = torch.randn((BATCH, H, len, len), dtype=dtype, device=device, requires_grad=True)
-
-    sm_scale = 1.0 / math.sqrt(HEAD_DIM)
-
-    def measure_max_error():
-        # For forward pass
-        if mode == "fwd":
-            out_triton = flash_attn_w_bias(q, k, v, bias, False, sm_scale)
-            out_torch = attention_torch(q, k, v, sm_scale, bias)
-            return torch.abs(out_triton - out_torch).max().item()
-        # For backward pass
-        else:
-            # Run triton backward
-            out_triton = flash_attn_w_bias(q, k, v, bias, False, sm_scale)
-            loss_triton = (0 - out_triton.sum())
-            loss_triton.backward()
-            dbias_triton = bias.grad.clone()
-            bias.grad = None
-
-            # Run torch backward
-            out_torch = attention_torch(q, k, v, sm_scale, bias)
-            loss_torch = (0 - out_torch.sum())
-            loss_torch.backward()
-            dbias_torch = bias.grad.clone()
-            bias.grad = None
-
-            # Compare gradients
-            return torch.abs(dbias_triton - dbias_torch).max().item()
-
-    def forward_backward_triton():
-        out = flash_attn_w_bias(q, k, v, bias, False, sm_scale)
-        loss = (0 - out.sum())
-        loss.backward()
-        torch.cuda.synchronize()
-        # Reset gradients for next iteration
-        q.grad = None
-        k.grad = None
-        v.grad = None
-        bias.grad = None
-
-    def forward_backward_torch():
-        out = attention_torch(q, k, v, sm_scale, bias)
-        loss = (0 - out.sum())
-        loss.backward()
-        torch.cuda.synchronize()
-        # Reset gradients for next iteration
-        q.grad = None
-        k.grad = None
-        v.grad = None
-        bias.grad = None
-
-    def forward_backward_sdpa():
-        out = attention_sdpa(q, k, v, sm_scale, bias)
-        loss = (0 - out.sum())
-        loss.backward()
-        torch.cuda.synchronize()
-        # Reset gradients for next iteration
-        q.grad = None
-        k.grad = None
-        v.grad = None
-        bias.grad = None
-
-    if mode == "fwd":
-        if provider == "triton-attn-bias":
-            fn = lambda: flash_attn_w_bias(q, k, v, bias, False, sm_scale)
-        elif provider == "torch-attn-bias":
-            fn = lambda: attention_torch(q, k, v, sm_scale, bias)
-        else:  # provider == "torch-sdpa"
-            fn = lambda: attention_sdpa(q, k, v, sm_scale, bias)
-    else:  # mode == "bwd"
-        if provider == "triton-attn-bias":
-            fn = forward_backward_triton
-        elif provider == "torch-attn-bias":
-            fn = forward_backward_torch
-        else:  # provider == "torch-sdpa"
-            fn = forward_backward_sdpa
-
-    # Run benchmark
-    ms = triton.testing.do_bench(fn)
-
-    # Calculate FLOPS
-    # Forward pass:
-    # FLOPS = 2 * QK + 2 * PV multiplications + bias addition
-    # QK matmul: 2 * B * H * N * N * D 
-    # PV matmul: 2 * B * H * N * N * D
-    # Bias addition: B * H * N * N
-    flops_qk = 2.0 * BATCH * H * len * len * HEAD_DIM  # QK matmul
-    flops_pv = 2.0 * BATCH * H * len * len * HEAD_DIM  # PV matmul 
-    flops_bias = BATCH * H * len * len  # Bias addition
-    total_flops_fwd = flops_qk + flops_pv + flops_bias
-
-    # Backward pass has roughly 2x the FLOPS of forward pass
-    total_flops = total_flops_fwd * (3 if mode == "bwd" else 1)
-
-    # Calculate TFLOPS
-    tflops = total_flops * 1e-12 / (ms * 1e-3)
-
-    # Measure max error (only for triton implementation)
-    max_error = measure_max_error() if provider == "triton-attn-bias" else 0.0
-
-    # Return TFLOPS for plotting
-    return tflops
-
-
-def test_attention_correctness():
-    """
-    Test the correctness of the Triton FlashAttention implementation against PyTorch.
-    Reports only the max absolute error for the forward pass and gradients.
-    """
-    torch.manual_seed(0)
-    device = torch.device("cuda")
-
-    # Fixed sequence length
-    batch, seqlen, nheads, d = 1, 2 ** 10, 2, 32
-    softmax_scale = 1.0 / math.sqrt(d)
-    causal = False
-
-    # Initialize input tensors
-    q, k, v = [
-        torch.randn(batch, seqlen, nheads, d, device=device, dtype=torch.float16, requires_grad=True)
-        for _ in range(3)
-    ]
-    bias = torch.randn(batch, nheads, seqlen, seqlen, device=device, dtype=torch.float16, requires_grad=True)
-
-    # Compute outputs
-    o_flash = flash_attn_w_bias(q, k, v, bias, causal, softmax_scale)
-    o_torch = attention_torch(q, k, v, softmax_scale, bias)
-    o_sdpa = attention_sdpa(q, k, v, softmax_scale, bias)
-
-    # Compute max forward errors
-    max_forward_error_torch = torch.abs(o_flash - o_torch).max().item()
-    max_forward_error_sdpa = torch.abs(o_flash - o_sdpa).max().item()
-
-    # Compute loss and backpropagate for flash attention
-    loss_flash = (0 - o_flash.sum())
-    loss_flash.backward()
-    dq_flash, dk_flash, dv_flash, dbias_flash = [x.grad.clone() for x in [q, k, v, bias]]
-
-    # Reset gradients
-    for x in [q, k, v, bias]:
-        x.grad = None
-
-    # Compute loss and backpropagate for torch attention
-    loss_torch = (0 - o_torch.sum())
-    loss_torch.backward()
-    dq_torch, dk_torch, dv_torch, dbias_torch = [x.grad.clone() for x in [q, k, v, bias]]
-
-    # Reset gradients
-    for x in [q, k, v, bias]:
-        x.grad = None
-
-    # Compute loss and backpropagate for SDPA
-    loss_sdpa = (0 - o_sdpa.sum())
-    loss_sdpa.backward()
-    dq_sdpa, dk_sdpa, dv_sdpa, dbias_sdpa = [x.grad.clone() for x in [q, k, v, bias]]
-
-    # Compute max gradient errors against torch implementation
-    max_grad_errors_torch = {
-        "q": torch.abs(dq_flash - dq_torch).max().item(),
-        "k": torch.abs(dk_flash - dk_torch).max().item(),
-        "v": torch.abs(dv_flash - dv_torch).max().item(),
-        "bias": torch.abs(dbias_flash - dbias_torch).max().item(),
-    }
-
-    # Compute max gradient errors against SDPA implementation
-    max_grad_errors_sdpa = {
-        "q": torch.abs(dq_flash - dq_sdpa).max().item(),
-        "k": torch.abs(dk_flash - dk_sdpa).max().item(),
-        "v": torch.abs(dv_flash - dv_sdpa).max().item(),
-        "bias": torch.abs(dbias_flash - dbias_sdpa).max().item(),
-    }
-
-    # Print max errors
-    print(f"✅ Max Forward Error vs Torch: {max_forward_error_torch:.6e}")
-    print(f"✅ Max Forward Error vs SDPA: {max_forward_error_sdpa:.6e}")
-    print(f"✅ Max Gradient Errors vs Torch:")
-    for param, err in max_grad_errors_torch.items():
-        print(f"   {param.upper()} : {err:.6e}")
-    print(f"✅ Max Gradient Errors vs SDPA:")
-    for param, err in max_grad_errors_sdpa.items():
-        print(f"   {param.upper()} : {err:.6e}")
-
-
-if __name__ == "__main__":
-    print("Running correctness tests...")
-    print("-" * 50)
-    test_attention_correctness()
-
-    print("\nRunning performance benchmarks...")
-    print("-" * 50)
-    df = bench_attention_comparison.run(save_path=".", print_data=True, return_df=True)
